examples/langchain/instrumentation-comparison/langfuse-sdk/pyproject.toml
[project]
name = "langchain-weather-langfuse"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "langchain-aws>=0.2.0",
    "langchain-core>=0.3.0",
    "langfuse>=3.0.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
